Extract Slack notification into notifiers/slack.py
