+++ tests/test_real_changesets.py
@@ -3,7 +3,7 @@
 import xml.etree.ElementTree as ET
 from pathlib import Path
 
-from watcher import detect_node_drags
+from checkers.drag import detect_node_drags
 
 FIXTURES = Path(__file__).parent / "fixtures"
